requirements.txt
requests
datetime
openai
io
docx-python